handlePopstateをPaginationに移動
